requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.6
django-cors-headers==4.3.1
django-ninja==1.3.0
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.11
pillow==12.0.0
protobuf==5.28.0
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
